// apps/api-auth/src/enums/auth-errors.enum.ts
export enum AuthErrorsEnum {
  INVALID_LOGIN_DATA = 'INVALID_LOGIN_DATA',
  MORE_DATA_NEEDED = 'MORE_DATA_NEEDED',
  INVALID_SINGUP_DATA = 'INVALID_SINGUP_DATA',
  EMAIL_ALREADY_EXIST = 'EMAIL_ALREADY_EXIST',
  EMAIL_NOT_VALID = 'EMAIL_NOT_VALID',
  NOT_AUTHORIZED = 'NOT_AUTHORIZED',
  TOKEN_NOT_VALID = 'TOKEN_NOT_VALID',
  INVALID_CODE = 'INVALID_CODE',
  USER_NOT_AVAILABLE_TO_COMPLETE = 'USER_NOT_AVAILABLE_TO_COMPLETE',
}